Fix 44 deep-review issues, add 4 missing extraction fields, update all docs
- Fix self-learning feedback loop (method tag mismatch, label uniqueness, accuracy gate)
- Fix security issues (tenant isolation always enforced, dev user=reviewer, credential validation)
- Fix database race conditions (atomic template activation, bulk deletes, thread-safe session)
- Fix extraction bugs (dead procedure_codes key, cross-field validators, confidence clamping)
- Add insurance_rep_name and insurance_rep_phone extraction fields (12 + 8 label aliases)
- Add payer_name and fax_received_date as SYSTEM-sourced metadata fields in API response
- Add 3 new database migrations (012-014: label unique constraint, active version index, bigint PKs)
- Add canonical CRITICAL_FIELDS constant (9 fields) as single source of truth
- Update all 14 documentation files to reflect new fields, sources, and extraction count (18 fields)

+++ libs/shared/db/models/fax_ocr_token.py
@@ -9,7 +9,7 @@ from decimal import Decimal
 from typing import TYPE_CHECKING
 from uuid import UUID
 
-from sqlalchemy import Boolean, ForeignKey, Integer, Numeric, Text
+from sqlalchemy import BigInteger, Boolean, ForeignKey, Integer, Numeric, Text
 from sqlalchemy.dialects.postgresql import UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -32,7 +32,7 @@ class FaxOcrToken(Base):
 
     # Primary key (BIGSERIAL for high volume)
     ocr_token_id: Mapped[int] = mapped_column(
-        Integer,
+        BigInteger,
         primary_key=True,
         autoincrement=True,
     )
